Package refactoring, introduced factory interfaces for services and stores creation
