+++ mobile/src/main/java/com/wedriveu/mobile/app/MainApplication.java
@@ -1,0 +1,7 @@
+package com.wedriveu.mobile.app;
+
+/**
+ * Created by nicolalasagni on 17/07/2017.
+ */
+public class MainApplication {
+}
