+++ mobile/src/main/java/com/wedriveu/mobile/service/ServiceFactory.java
@@ -1,0 +1,7 @@
+package com.wedriveu.mobile.service;
+
+/**
+ * Created by nicolalasagni on 17/07/2017.
+ */
+public interface ServiceFactory {
+}
