+++ mobile/src/main/java/com/wedriveu/mobile/store/StoreFactory.java
@@ -1,0 +1,7 @@
+package com.wedriveu.mobile.store;
+
+/**
+ * Created by nicolalasagni on 17/07/2017.
+ */
+public interface StoreFactory {
+}
